✨ feat: 모달 종류에 따라 mutate 처리 및 컴포넌트 분리
