+++ packages/frontend/src/pages/stock-detail/components/index.ts
@@ -2,3 +2,4 @@ export * from './TextArea';
 export * from './MetricItem';
 export * from './MetricSection';
 export * from './Title';
+export * from './StockDetailHeader';
